makes rendering time faster.
Previous implementation uses a scatter plot to present state data. This required to app to plot a huge amount of data even for a small dataset. This was causing the app to be slow and unresponsive. The new implementation only plots state transition. A line between two points is used to represent a node retaining its current state. This makes rendering time much faster.

Because of the fast render time, the hide state option has also been removed.